handle one to one relationship for rockets and height entity, implement saving and query process

+++ app/src/main/java/com/example/rockets/data/remote/dto/Engines.kt
@@ -8,7 +8,7 @@ data class Engines(
     val propellant_1: String,
     val propellant_2: String,
     val thrust_sea_level: ThrustSeaLevel,
-    val thrust_to_weight: Int,
+    val thrust_to_weight: Double,
     val thrust_vacuum: ThrustVacuum,
     val type: String,
     val version: String
