+++ app/src/main/java/com/example/rockets/data/remote/dto/FirstStage.kt
@@ -2,7 +2,7 @@ package com.example.rockets.data.remote.dto
 
 data class FirstStage(
     val burn_time_sec: Int,
-    val engines: Int,
+    val engines: Double,
     val fuel_amount_tons: Double,
     val reusable: Boolean,
 )
